Move Attention and MLP classes to a separate file in Megatron transformers (#5453)
* Move attention and mlp to separate files


* Add new attention and mlp files


* Fix import in tests


* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Remove unused imports in attention


* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix missing import


* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Fix
