+++ tests/collections/nlp/test_retrieval_module_inference.py
@@ -19,6 +19,7 @@ import torch.nn.functional as F
 from einops import rearrange
 from pytorch_lightning.trainer.trainer import Trainer
 
+from nemo.collections.nlp.modules.common.megatron.attention import ParallelChunkedCrossAttention
 from nemo.collections.nlp.modules.common.megatron.layer_type import LayerType
 from nemo.collections.nlp.modules.common.megatron.megatron_init import initialize_model_parallel_for_nemo
 from nemo.collections.nlp.modules.common.megatron.retrieval_token_level_encoder_decoder import (
@@ -29,7 +30,6 @@ from nemo.collections.nlp.modules.common.megatron.retrieval_transformer import (
     MegatronRetrievalTransformerEncoderModule,
 )
 from nemo.collections.nlp.modules.common.megatron.rotary_pos_embedding import RotaryEmbedding
-from nemo.collections.nlp.modules.common.megatron.transformer import ParallelChunkedCrossAttention
 from nemo.collections.nlp.modules.common.megatron.utils import (
     build_attention_mask_3d,
     init_method_normal,
